# front/openapi/family-agenda.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agenda familial - API",
        "description": "Un **Agenda** contient des **AgendaEntry** persist\u00e9es. Cr\u00e9er un `Reminder` g\u00e9n\u00e8re imm\u00e9diatement ses entr\u00e9es (une seule s'il est ponctuel, toute la fen\u00eatre glissante s'il est r\u00e9current) ; un job nocturne prolonge ensuite les r\u00e9currences.\n\nCommencer par la section **Agenda** : `GET /v1/agenda`.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/v1/agenda/entries/{id}": {
            "patch": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Changer le statut d'une occurrence",
                "description": "Marque UNE entr\u00e9e comme compl\u00e9t\u00e9e, annul\u00e9e ou d\u00e9plac\u00e9e, sans toucher aux autres occurrences du reminder.",
                "operationId": "updateEntryStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant de l'AgendaEntry",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgendaEntryStatusRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Entr\u00e9e introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgendaEntryDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/agenda": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Entr\u00e9es de l'agenda sur une plage de dates",
                "description": "Liste chronologique des AgendaEntry dont la date tombe entre `dateDebut` et `dateFin` (jours inclus).",
                "operationId": "listAgenda",
                "parameters": [
                    {
                        "name": "dateDebut",
                        "in": "query",
                        "description": "Premier jour (ISO, inclus)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-01"
                    },
                    {
                        "name": "dateFin",
                        "in": "query",
                        "description": "Dernier jour (ISO, inclus)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-30"
                    },
                    {
                        "name": "membreId",
                        "in": "query",
                        "description": "Ne garder que les entr\u00e9es concernant ce membre (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "400": {
                        "description": "dateFin ant\u00e9rieure \u00e0 dateDebut, ou date mal form\u00e9e",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AgendaEntryDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/agenda/semaine/{date}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Entr\u00e9es de la semaine d'une date",
                "description": "Entr\u00e9es de la semaine ISO (lundi \u00e0 dimanche) qui contient `date`, en ordre chronologique.",
                "operationId": "getAgendaSemaine",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "N'importe quel jour de la semaine (ISO)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-23"
                    },
                    {
                        "name": "membreId",
                        "in": "query",
                        "description": "Ne garder que les entr\u00e9es concernant ce membre (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AgendaEntryDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/agenda/mois/{annee}/{mois}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Entr\u00e9es d'un mois, group\u00e9es par jour",
                "description": "Map `date -> entr\u00e9es`. Seuls les jours ayant au moins une entr\u00e9e sont pr\u00e9sents.",
                "operationId": "getAgendaMois",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "maximum": 2200,
                            "minimum": 1900
                        },
                        "example": 2026
                    },
                    {
                        "name": "mois",
                        "in": "path",
                        "description": "Mois (1-12)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "maximum": 12,
                            "minimum": 1
                        },
                        "example": 9
                    },
                    {
                        "name": "membreId",
                        "in": "query",
                        "description": "Ne garder que les entr\u00e9es concernant ce membre (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "2026-09-22": [
                                        {
                                            "id": 1,
                                            "reminderId": 1,
                                            "titre": "Football",
                                            "type": "SPORT",
                                            "dateHeure": "2026-09-22T18:00:00",
                                            "dateHeureFin": "2026-09-22T19:30:00",
                                            "statut": "PREVU",
                                            "membres": [
                                                {
                                                    "id": 4,
                                                    "nom": "Tom",
                                                    "couleur": "#4CAF50"
                                                }
                                            ],
                                            "recurring": true
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/agenda/jour/{date}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Entr\u00e9es d\u00e9taill\u00e9es d'un jour",
                "operationId": "getAgendaJour",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Jour (ISO)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2026-09-22"
                    },
                    {
                        "name": "membreId",
                        "in": "query",
                        "description": "Ne garder que les entr\u00e9es concernant ce membre (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AgendaEntryDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/agenda/annee/{annee}": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "R\u00e9sum\u00e9 annuel",
                "description": "Pour chacun des 12 mois : nombre d'entr\u00e9es, au total et par type de reminder. Sert \u00e0 la vue annuelle.",
                "operationId": "getAgendaAnnee",
                "parameters": [
                    {
                        "name": "annee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "maximum": 2200,
                            "minimum": 1900
                        },
                        "example": 2026
                    },
                    {
                        "name": "membreId",
                        "in": "query",
                        "description": "Ne garder que les entr\u00e9es concernant ce membre (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/YearSummaryDTO"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "FamilyMemberDTO": {
                "type": "object",
                "description": "Membre de la famille",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "L\u00e9a"
                    },
                    "couleur": {
                        "type": "string",
                        "example": "#E91E63"
                    }
                }
            },
...
            "AgendaEntryStatusRequest": {
                "type": "object",
                "description": "Nouveau statut d'une occurrence",
                "example": {
                    "statut": "ANNULE"
                },
                "properties": {
                    "statut": {
                        "type": "string",
                        "description": "PREVU, COMPLETE, ANNULE ou DEPLACE",
                        "enum": [
                            "PREVU",
                            "COMPLETE",
                            "ANNULE",
                            "DEPLACE"
                        ],
                        "example": "ANNULE"
                    }
                },
                "required": [
                    "statut"
                ]
            },
            "AgendaEntryDTO": {
                "type": "object",
                "description": "Occurrence d'un reminder dans l'agenda",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Identifiant de l'entr\u00e9e (\u00e0 utiliser pour le PATCH de statut)",
                        "example": 42
                    },
                    "reminderId": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "titre": {
                        "type": "string",
                        "example": "Football"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SPORT",
                            "TEST_MEDICAL",
                            "RENDEZ_VOUS",
                            "AUTRE"
                        ],
                        "example": "SPORT"
                    },
                    "dateHeure": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Date/heure de cette occurrence",
                        "example": "2026-09-22T18:00:00"
                    },
                    "dateHeureFin": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Fin de cette occurrence (null si le reminder n'a pas de fin)",
                        "example": "2026-09-22T19:30:00"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "PREVU",
                            "COMPLETE",
                            "ANNULE",
                            "DEPLACE"
                        ],
                        "example": "PREVU"
                    },
                    "membres": {
                        "type": "array",
                        "description": "Membres concern\u00e9s",
                        "items": {
                            "$ref": "#/components/schemas/FamilyMemberDTO"
                        }
                    },
                    "recurring": {
                        "type": "boolean",
                        "description": "L'entr\u00e9e provient d'un reminder r\u00e9current"
                    }
                }
            },
            "MonthSummaryDTO": {
                "type": "object",
                "description": "R\u00e9sum\u00e9 d'un mois",
                "properties": {
                    "mois": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Num\u00e9ro du mois (1-12)",
                        "example": 9
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Nombre total d'entr\u00e9es du mois",
                        "example": 9
                    },
                    "parType": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "description": "Nombre d'entr\u00e9es par type (types absents = 0)",
                        "example": {
                            "SPORT": 8,
                            "TEST_MEDICAL": 1
                        }
                    }
                }
            },
            "YearSummaryDTO": {
                "type": "object",
                "description": "Vue annuelle : nombre d'entr\u00e9es par mois, group\u00e9es par type",
                "properties": {
                    "annee": {
                        "type": "integer",
                        "format": "int32",
                        "example": 2026
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Nombre total d'entr\u00e9es de l'ann\u00e9e",
                        "example": 42
                    },
                    "mois": {
                        "type": "array",
                        "description": "12 \u00e9l\u00e9ments, de janvier \u00e0 d\u00e9cembre",
                        "items": {
                            "$ref": "#/components/schemas/MonthSummaryDTO"
                        }
                    }
                }
            }
        }
    }
}