# front/src/app/api-client/model/agenda-entry-status-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Nouveau statut d\'une occurrence
 */
export interface AgendaEntryStatusRequest { 
    /**
     * PREVU, COMPLETE, ANNULE ou DEPLACE
     */
    statut: AgendaEntryStatusRequest.StatutEnum;
}
export namespace AgendaEntryStatusRequest {
    export const StatutEnum = {
        Prevu: 'PREVU',
        Complete: 'COMPLETE',
        Annule: 'ANNULE',
        Deplace: 'DEPLACE',
    } as const;
    export type StatutEnum = typeof StatutEnum[keyof typeof StatutEnum];
}